restructure bin/drk as a library
